backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.35
feedparser==6.0.11
requests==2.31.0
pillow>=10.3.0
python-multipart==0.0.6
aiofiles==23.2.1
apscheduler==3.10.4
